+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Bumped vitest and playwright packages
 - Bumped base image to debian:13.5
 - Fixed the issue that a closed release is still shown in the dropdown for releases in the board filter and the issuer modal
+- Allow single spaces in configurable column/status names
 
 # 1.3.0